refactor: Separate server from app file (integration. tests)
